+++ NEWS.md
@@ -1,7 +1,7 @@
 # encryptr 0.1.3
 
 * File encryption include, `encrypt_file()` and `decrypt_file()`. 
-
+* But in `encrypt()` and `encrypt_file()` fixed for public key at URL. #5 
 
 # encryptr 0.1.2